Applied patch #2832758: p2os big patch

+++ server/drivers/mixed/p2os/robot_params.h
@@ -190,7 +190,7 @@ typedef struct
   int Qtt; //
   int Qxx; //
   int Qyy; //
-  int RangeConvFactor; //
+  double RangeConvFactor; //
   int RearBumpers; //
   int RecoverOnFail; //
   int ReflectorMatchDist; //
